Add seqchecker
Chronicle is expected to be sequentially consistent when performing
sticky non-quorum reads. Our sequential consistency checker from the
jepsen.couchbase repo is useful in verifying this. Eventually it might
be better to split the common components of jepsen.couchbase and
jepsen.chronicle into a shared repo that both depend on, but for now
just copy over the sequential consistency checker.